Create all command classes in command package
