+++ src/main/java/commands/Command.java
@@ -2,6 +2,6 @@ package commands;
 import DukeException.DukeException;
 import tasks.TaskList;
 
-public interface Command {
-    void execute(String details, TaskList tasks) throws DukeException;
+public abstract class Command {
+    public void execute(TaskList tasks) throws DukeException {}
 }
