+++ src/main/java/tasks/Deadline.java
@@ -1,14 +1,14 @@
-import tasks.Task;
+package tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.io.Serializable;
 
 public class Deadline extends Task implements Serializable {
-    private LocalDateTime deadline;
+    private final LocalDateTime deadline;
     private static final long serialVersionUID = 3L;
 
-    Deadline(String description, LocalDateTime deadline) {
+    public Deadline(String description, LocalDateTime deadline) {
         super(description);
         this.deadline = deadline;
     }
@@ -16,6 +16,6 @@ public class Deadline extends Task implements Serializable {
     @Override
     public String toString() {
         return "[D]" + "[" + (isDone ? "X" : " ") + "] " + description + "(by: "
-                + deadline.format(DateTimeFormatter.ofPattern("HH:mm MMM dd YYYY")) + ")";
+                + deadline.format(DateTimeFormatter.ofPattern("HH:mm MMM dd yyyy")) + ")";
     }
 }
